Draft y-level-lock + other fixes
Camera now set to orthographic for clearer gameplay and aesthetics
GameManager now holds a float (should be an int actually - to fix) that determines the y-level of the current box, and methods to increase and decrease currently bound to up and down arrows
Box movement now along a plane, which takes a raycast from the mouse. This plane is updated along with the y-float.
Directional light altered to give shadows accurate to placement, for better UX (to be replaced with blueprint shadows)
Placeholder grid instantiated at runtime, though currently the  boxes move in half-steps along X and Z rather than full-steps